requirements.txt
django-unfold==0.49.*
django-svelte-jsoneditor==0.4.*
django-import-export==4.0.*
django-admin-confirm==1.0.*
django-mass-edit==3.5.*
django-cleanup==9.0.*
